Quizzes can be Created
Quizzes can be created and added to the database without questions. We need Quiz_ID in the QUESTIONS table of the database to link the questions, but once that is set up, quizzes can be created properly.
These Quizzes cannot be viewed yet since we need to have a python file that generates a quiz page.